pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "quill"
version = "0.1.0"
description = "A command-line interface tool for interacting with Google Drive"
readme = "README.md"
requires-python = ">=3.9"
license = "MIT"
authors = [
    { name = "Ignasi Fosch", email = "natx@y10k.ws" }
]
dependencies = [
    "click>=8.1.7",
    "google-api-python-client>=2.108.0",
    "google-auth-httplib2>=0.1.1",
    "google-auth-oauthlib>=1.1.0",
    "pytest>=8.4.0",
    "pytest-cov>=6.2.1",
    "python-dateutil>=2.9.0.post0",
]

[project.optional-dependencies]
dev = [
    "myst-parser>=3.0.1",
    "pre-commit>=3.6.0",
    "pytest>=8.3.5",
    "pytest-cov>=5.0.0",
    "ruff>=0.11.13",
    "sphinx>=7.1.2",
    "sphinx-rtd-theme>=3.0.2",
    "ty>=0.0.1a9",
]

[project.scripts]
quill = "quill.cli:cli"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=quill --cov-report=term-missing --cov-fail-under=80"

[tool.ruff]
line-length = 88
target-version = "py38"

[dependency-groups]
dev = [
    "ruff>=0.11.13",
    "ty>=0.0.1a9",
]
